+++ src/utils/combineActions.ts
@@ -1,0 +1,8 @@
+export default (...actions) => (...actionsParams) =>
+  actions.reduce(
+    (acc, action) => ({
+      ...acc,
+      ...(typeof action === "function" ? action(...actionsParams) : action)
+    }),
+    {}
+  );